+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - `esbuild-serve init` to generate an example index.html file.
+- `esbuild-serve --build`.
 
 ## [0.3.4] - 2021-06-15
 Breaking: This version completely rewrites this library. Now it becomes a tiny